+++ src/Components/index.js
@@ -1,5 +1,5 @@
-export { default as Footer } from "./Layout/Footer"
-export { default as Header } from "./Layout/Header"
-export { default as Layout } from "./Layout/Layout"
-export { default as Navbar } from "./Layout/Navbar"
-export * from "./index"
+export { default as Footer } from "./Layout/Footer";
+export { default as Header } from "./Layout/Header/Header";
+export { default as Layout } from "./Layout/Layout";
+export { default as Navbar } from "./Layout/Navbar";
+export * from "./index";
